Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
     int r=1;
-    int Area=(3.14*r)*r;
-    scanf("%d",&Area);
-    printf("Area %d",Area);
+    float Area=(3.14*r)*r;
+    scanf("%f",&Area);
+    printf("Area %.2f",Area);
     return 0;
 }
